Refactor ModalSheet component and ModalSheetProvider component, and add types.d.ts file
