+++ package.json
@@ -3,6 +3,7 @@
   "version": "1.4.2",
   "description": "Modal sheet component for React Native. Inspired by the native modal sheet in iOS.",
   "main": "src/index.ts",
+  "types": "src/types.d.ts",
   "scripts": {
     "open:ios": "open -a \"Xcode\" example/ios",
     "open:android": "open -a \"Android Studio\" example/android"
